% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnparser.R
\name{flatten_result}
\alias{flatten_result}
\title{Parse names using gnparser}
\usage{
flatten_result(json)
}
\arguments{
\item{json}{JSON string from the gnparser library}
}
\value{
A data.frame
}
\description{
Parse names using gnparser
}
