% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnparser.R
\docType{package}
\name{gnparser}
\alias{gnparser-package}
\alias{gnparser}
\title{gnparser}
\description{
R wrapper for gnparser Go library
}
\author{
\strong{Maintainer}: Pieter Provoost \email{pieterprovoost@gmail.com}

}
