% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnparser.R
\docType{package}
\name{gnparser}
\alias{gnparser-package}
\alias{gnparser}
\title{gnparser}
\description{
R interface to the gnparser library for parsing scientific names.
}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
